+++ peli-xr.js
@@ -19,7 +19,7 @@
  
 (function(plugin) {
 
-// var version = '0.10.0b';
+// var version = '0.10.1b';
 
 
 //////////////////////////////////////////////////////////////////////////
@@ -390,123 +390,6 @@
 
 	}
 	//HostFactory.registrarHost("filenuke",Filenuke); //Registrar la clase Filenuke
-
-	/********************************************************************************	
-	/* var Magnovideo: Objeto que representa el servidor Magnovideo					*
-	/*		Hereda de Host															*
-	/********************************************************************************/
-	var Magnovideo= function() { //SERVER CERRADo
-
-		//metodos publicos
-
-		/************************************************************************
-		/*	funcion esservidoradulto: Indica si es un servidor de adultos o no. *
-		/*	Parametros: ninguno													*
-		/*	Retorna: true si es un servidor de adultos, false si no lo es.		*
-		/***********************************************************************/
-		this.esservidoradulto= function () {
-			return false;
-		}
-
-		/****************************************************************************
-		/*	funcion geturl_video: Devuelve la url del video.						*
-		/*	Parametros:																*
-		/*		url_servidor: direccion de la que se debe extraer la url del video.	*	
-		/*	Retorna: String que representa la url del video o 'error'				*									*
-		/***************************************************************************/
-		this.geturl_video= function (url_servidor)
-		{
-			var id_magnovideo;
-			id_magnovideo = url_servidor.replace('http://www.magnovideo.com/?v=','');
-			url_servidor = 'http://www.magnovideo.com/player_config.php?mdid=' + id_magnovideo + '&sml=1&autoplay=true';
-
-			var file_contents = get_urlsource(url_servidor);	
-			var storage_path = extraer_texto(file_contents,'<storage_path>','</storage_path>');
-			var video_name = extraer_texto(file_contents,'<video_name>','</video_name>');
-			var title_thumbs = extraer_texto(file_contents,'<tile_thumbs>','</tile_thumbs>');
-			var movie_burst  = extraer_texto(file_contents,'<movie_burst>','</movie_burst>') + 'k';
-			var burst_speed  = extraer_texto(file_contents,'<burst_speed>','</burst_speed>') + 'k';
-			var ste  = extraer_texto(file_contents,'<ste>','</ste>');
-
-			//construir la url
-			var pos_aux;
-			var aux_string = title_thumbs;
-
-			pos_aux = aux_string .indexOf('storage/files/');
-			aux_string = aux_string .substr(pos_aux+14);
-			pos_aux = aux_string .indexOf('/');
-			aux_string = aux_string .substr(pos_aux+1);
-			aux_string = aux_string.replace('tmpsmall/tiles.jpg','');
-
-			var url_video = storage_path + 'storage/files/' + aux_string + video_name + '?burst=' + movie_burst + '&u=' + burst_speed + '&' + ste;
-
-			showtime.sleep(1000);
-
-			//showtime.trace('Debug: ' + url_video);
-			//"http://e1.magnovideo.com:8080/storage/files/0/15/107/52/1.mp4?burst=5905k&u=600k&md=3LOQPhshwrZSW1knDdUGiA&e=1391099841"
-			//http://www.magnovideo.com/player_config.php?mdid="+id_video+"&sml=1&autoplay=true
-			//http://www.magnovideo.com/?v=JCRFATY
-			//http://www.magnovideo.com/player_config.php?mdid=JCRFATY&sml=1&autoplay=true
-
-		return url_video;
-		}
-
-	}
-	//HostFactory.registrarHost("magnovideo",Magnovideo); //Registrar la clase Magnovideo
-
-	/********************************************************************************	
-	/* var Mightyupload: Objeto que representa el servidor Mightyupload				*
-	/*		Hereda de Host															*
-	/*******************************************************************************/
-	var Mightyupload= function() {
-
-		//metodos publicos
-
-		/************************************************************************
-		/*	funcion esservidoradulto: Indica si es un servidor de adultos o no. *
-		/*	Parametros: ninguno													*
-		/*	Retorna: true si es un servidor de adultos, false si no lo es.		*
-		/***********************************************************************/
-		this.esservidoradulto= function () {
-			return false;
-		}
-
-		/****************************************************************************
-		/*	funcion geturl_video: Devuelve la url del video.						*
-		/*	Parametros:																*
-		/*		url_servidor: direccion de la que se debe extraer la url del video.	*	
-		/*	Retorna: String que representa la url del video o 'error'				*									*
-		/***************************************************************************/
-		this.geturl_video= function (url_servidor)
-		{
-			var file_contents = get_urlsource(url_servidor);
-			var error = file_contents.indexOf('File Not Found')
-			if(error==-1)
-			{
-				var op = extraer_texto(file_contents,'<input type="hidden" name="op" value="','"');
-				var id = extraer_texto(file_contents,'<input type="hidden" name="id" value="','"');
-				var rand = extraer_texto(file_contents,'<input type="hidden" name="rand" value="','"');
-				var referer = extraer_texto(file_contents,'<input type="hidden" name="referer" value="','"');
-				var plugins_are_not_allowed = extraer_texto(file_contents,'<input type="hidden" name="plugins_are_not_allowed" value="','"');
-				var method_free = extraer_texto(file_contents,'<input type="hidden" name="method_free" value="','"');
-				var method_premium = extraer_texto(file_contents,'<input type="hidden" name="method_premium" value="','"');
-				var down_direct = extraer_texto(file_contents,'<input type="hidden" name="down_direct" value="','"');
-				file_contents = ""
-				showtime.sleep(1000);
-
-				var datos_post = {'op':op, 'id':id, 'rand':rand, 'referer':referer, 'plugins_are_not_allowed':plugins_are_not_allowed, 'method_free':method_free, 'method_premium':method_premium, 'down_direct':down_direct};
-				file_contents = post_urlsource(url_servidor,datos_post);
-				var url_video = extraer_texto(file_contents,"file: '","'");
-				file_contents = "";
-			}else{
-				url_video='error';
-			}
-		return url_video;
-		}
-
-
-	}
-	//HostFactory.registrarHost("mightyupload",Mightyupload); //Registrar la clase Mightyupload
 
 	/********************************************************************************	
 	/* var Movshare: Objeto que representa el servidor Movshare						*
@@ -1446,11 +1329,10 @@
 		this.getmenu= function(page){
 		//retorna el Menu
 			var array_menu=[
-				//new Item_menu('Br-Dvd Castellano','views/img/folder.png',':vercontenido:newpct:tipobrdvd:'+ escape('http://www.newpct1.com/peliculas-castellano/peliculas-rip/')),
-				new Item_menu('Br-Dvd Castellano','views/img/folder.png',':vercontenido:newpct:tipobrdvd:'+ escape('http://www.newpct1.com/peliculas/')),
-				new Item_menu('Estrenos de Cine','views/img/folder.png',':vercontenido:newpct:tipoestrenoscine:' + escape('http://www.newpct1.com/peliculas-castellano/estrenos-de-cine/;1')),
-				new Item_menu('V.O. Subtituladas','views/img/folder.png',':vercontenido:newpct:tipovo:' + escape('http://www.newpct1.com/peliculas-vo/;1')),
-				new Item_menu('Buscar','views/img/search.png',':vercontenido:newpct:tipobusqueda:' + escape('http://www.newpct1.com/buscar-descargas/'))
+				new Item_menu('Br-Dvd Castellano','views/img/folder.png',':vercontenido:newpct:tipobrdvd:'+ escape('http://www.newpct1.com/index.php?page=categorias&url=peliculas&letter=&pg=1')),
+				new Item_menu('Estrenos de Cine','views/img/folder.png',':vercontenido:newpct:tipoestrenoscine:' + escape('http://www.newpct1.com/index.php?page=categorias&url=estrenos-de-cine&letter=&pg=1')),
+				new Item_menu('V.O. Subtituladas','views/img/folder.png',':vercontenido:newpct:tipovo:' + escape('http://www.newpct1.com/index.php?page=categorias&url=peliculas-vo&letter=&pg=1')),
+				new Item_menu('Buscar','views/img/search.png',':vercontenido:newpct:tipobusqueda:' + escape('http://www.newpct1.com/buscar'))
 			];
 
 			return array_menu;
@@ -1472,34 +1354,33 @@
 		switch (tipo)
 			{
 			case "tipobrdvd":
+			
 				page.metadata.title = "NewPct - Br-Dvd Castellano";
-				var url_2=escape('http://www.newpct1.com/include.inc/ajax.php/orderCategory.php?type=todo&leter=&sql=SELECT+DISTINCT+++%09%09%09%09%09%09torrentID%2C+++%09%09%09%09%09%09torrentCategoryID%2C+++%09%09%09%09%09%09torrentCategoryIDR%2C+++%09%09%09%09%09%09torrentImageID%2C+++%09%09%09%09%09%09torrentName%2C+++%09%09%09%09%09%09guid%2C+++%09%09%09%09%09%09torrentShortName%2C++%09%09%09%09%09%09torrentLanguage%2C++%09%09%09%09%09%09torrentSize%2C++%09%09%09%09%09%09calidad+as+calidad_%2C++%09%09%09%09%09%09torrentDescription%2C++%09%09%09%09%09%09torrentViews%2C++%09%09%09%09%09%09rating%2C++%09%09%09%09%09%09n_votos%2C++%09%09%09%09%09%09vistas_hoy%2C++%09%09%09%09%09%09vistas_ayer%2C++%09%09%09%09%09%09vistas_semana%2C++%09%09%09%09%09%09vistas_mes++%09%09%09%09++FROM+torrentsFiles+as+t+WHERE++(torrentStatus+%3D+1+OR+torrentStatus+%3D+2)++AND+(torrentCategoryID+IN+(1537%2C+758%2C+1105%2C+760%2C+1225))++AND+home_active+%3D+0++++ORDER+BY+torrentDateAdded++DESC++LIMIT+0%2C+50&pag=1&tot=&ban=3&cate=1225');
-				var params={'url_servidor': unescape(url_2),
+				var params={'url_servidor': unescape(url),
 					'page_uri': ':verenlaces:newpct:',
 					'uri_siguiente': ':vercontenido:newpct:tipo1:',
 					'subtitulo':false}	
-				array_playlist=this.parsenewpcttipodefault(url,params);
+				array_playlist= this.parsenewpct (params);
 				break;
 			case "tipoestrenoscine":
 				page.metadata.title = "NewPct - Estrenos de Cine";
-				var url_2=escape('http://www.newpct1.com/include.inc/ajax.php/orderCategory.php?type=todo&leter=&sql=SELECT+DISTINCT+++%09%09%09%09%09%09torrentID%2C+++%09%09%09%09%09%09torrentCategoryID%2C+++%09%09%09%09%09%09torrentCategoryIDR%2C+++%09%09%09%09%09%09torrentImageID%2C+++%09%09%09%09%09%09torrentName%2C+++%09%09%09%09%09%09guid%2C+++%09%09%09%09%09%09torrentShortName%2C++%09%09%09%09%09%09torrentLanguage%2C++%09%09%09%09%09%09torrentSize%2C++%09%09%09%09%09%09calidad+as+calidad_%2C++%09%09%09%09%09%09torrentDescription%2C++%09%09%09%09%09%09torrentViews%2C++%09%09%09%09%09%09rating%2C++%09%09%09%09%09%09n_votos%2C++%09%09%09%09%09%09vistas_hoy%2C++%09%09%09%09%09%09vistas_ayer%2C++%09%09%09%09%09%09vistas_semana%2C++%09%09%09%09%09%09vistas_mes++%09%09%09%09++FROM+torrentsFiles+as+t+WHERE++(torrentStatus+%3D+1+OR+torrentStatus+%3D+2)++AND+(torrentCategoryID+IN+(1231%2C+1165%2C+1230%2C+1232%2C+766%2C+765%2C+761%2C+848%2C+1224))++AND+home_active+%3D+0++++ORDER+BY+torrentDateAdded++DESC++LIMIT+0%2C+50&pag=1&tot=&ban=3&cate=1224');
-				var params={'url_servidor': unescape(url_2),
+				//var url_2='http://www.newpct1.com/index.php?page=categorias&url=estrenos-de-cine&letter=&pg=1';
+				var params={'url_servidor': unescape(url),
 					'page_uri': ':verenlaces:newpct:',
 					'uri_siguiente': ':vercontenido:newpct:tipo1:',
 					'subtitulo':false}	
-				array_playlist=this.parsenewpcttipodefault(url,params);
+				array_playlist= this.parsenewpct (params);
 				break;
 			case "tipovo":
 				page.metadata.title = "NewPct - V.O. Subtituladas";
-				var url_2=escape('http://www.newpct1.com/include.inc/ajax.php/orderCategory.php?type=todo&leter=&sql=SELECT+DISTINCT+++%09%09%09%09%09%09torrentID%2C+++%09%09%09%09%09%09torrentCategoryID%2C+++%09%09%09%09%09%09torrentCategoryIDR%2C+++%09%09%09%09%09%09torrentImageID%2C+++%09%09%09%09%09%09torrentName%2C+++%09%09%09%09%09%09guid%2C+++%09%09%09%09%09%09torrentShortName%2C++%09%09%09%09%09%09torrentLanguage%2C++%09%09%09%09%09%09torrentSize%2C++%09%09%09%09%09%09calidad+as+calidad_%2C++%09%09%09%09%09%09torrentDescription%2C++%09%09%09%09%09%09torrentViews%2C++%09%09%09%09%09%09rating%2C++%09%09%09%09%09%09n_votos%2C++%09%09%09%09%09%09vistas_hoy%2C++%09%09%09%09%09%09vistas_ayer%2C++%09%09%09%09%09%09vistas_semana%2C++%09%09%09%09%09%09vistas_mes++%09%09%09%09++FROM+torrentsFiles+as+t+WHERE++(torrentStatus+%3D+1+OR+torrentStatus+%3D+2)++AND+(torrentCategoryID+IN+(779%2C+784%2C+787%2C+788%2C+786%2C+778))++AND+home_active+%3D+0++++ORDER+BY+torrentDateAdded++DESC++LIMIT+0%2C+50&pag=1&tot=&ban=3&cate=778');
-				var params={'url_servidor': unescape(url_2),
+				//var url_2='http://www.newpct1.com/index.php?page=categorias&url=peliculas-vo&letter=&pg=1';
+				var params={'url_servidor': unescape(url),
 					'page_uri': ':verenlaces:newpct:',
 					'uri_siguiente': ':vercontenido:newpct:tipo1:',
 					'subtitulo':false}	
-				array_playlist=this.parsenewpcttipodefault(url,params);
+				array_playlist= this.parsenewpct (params);
 				break;
 			case "tipo1":
-				//array_playlist=parsenewpctpeliculastipo1(page,url);
 				var params={'url_servidor': unescape(url),
 					'page_uri': ':verenlaces:newpct:',
 					'uri_siguiente': ':vercontenido:newpct:tipo1:',
@@ -1537,39 +1418,30 @@
 		var calidad;
 		var descripcion;
 
-		//check login
-
-		file_contents = that.checkloginnewpct(url, file_contents);
+		
 		if(file_contents!=false)
 			{
 			//item_Actual
-			//titulo = extraer_texto(file_contents ,'<img itemprop="image"','/>');
-			//titulo = extraer_texto(titulo ,'alt="','"');
-			titulo = extraer_texto(file_contents ,'<title>','Descargar');
-			imagen = extraer_texto(file_contents ,'<meta property="og:image" content="','" />');
-			descripcion = extraer_texto(file_contents ,"</div><div class='sinopsis'>","</div>");
+			titulo = extraer_texto(file_contents ,'<h1><strong>','</strong>');
+			imagen = extraer_texto(file_contents ,'<div class="entry-left">','</div>');
+			imagen = extraer_texto(imagen ,'<img src="','"');
+			descripcion = extraer_texto(file_contents ,"<div class='descripcion_top'>","</div>");
 
 			this.item_Actual=new Item_menu(titulo,imagen,null,url,descripcion);
 
-
-			file_contents = extraer_texto(file_contents,"<thead id='ver-online'>","</table>");
-			file_contents = extraer_texto(file_contents,'<tbody>','</tbody>');
-			//<tr to </tr>
-			var array_aux = extraer_html_array(file_contents,'<tr','</tr>');
+			file_contents = extraer_texto(file_contents,'<div id="tab3"','</ul>');
+			var array_aux = extraer_html_array(file_contents,'<div class="box1">','<div class="box6">');
 			file_contents = "";
 
 
 			var array_aux2=[];
 			for (var i=0;i<array_aux.length;i++)
 				{
-				if(array_aux[i].indexOf('<td>ver en 1 Link</td>')!='-1')
-					{
+					servidor = extraer_texto(array_aux[i],'<div class="box2">','</div>');	
+					idioma = extraer_texto(array_aux[i],'<div class="box3">','</div>');
+					calidad = extraer_texto(array_aux[i],'<div class="box4">','</div>');
 					url_host = extraer_texto(array_aux[i],"<a href='","'");
-					array_aux2 = extraer_html_array(array_aux[i],'<td>','</td>');
-					servidor = extraer_texto(array_aux2[0],'<td>','</td>');	
-					idioma = extraer_texto(array_aux2[1],'<td>','</td>');
-					calidad = extraer_texto(array_aux2[2],'<td>','</td>');
-
+				
 					var params={
 							"url_host" : url_host,
 							"servidor" : servidor,
@@ -1582,9 +1454,7 @@
 						{ 
 							array_servidores.push(objHost);
 						} 
-					}
 				}
-
 			}
 		return array_servidores;
 		}
@@ -1601,72 +1471,23 @@
 		}
 
 
-		//Estos metodos son publicos para poder acceder desde los herederos
-		this.checkloginnewpct= function (url_servidor, file_contents)
-		{
-		var logueado = file_contents.indexOf("//loadtopBar('');")
-		var valor_retorno=false;
-		if(logueado!=-1) //no logueado
-			{
-			var reason = 'Introduce tu usuario y contraseña de http://www.newpct.com';
-			var do_query = false;
-			while(1)
-				{
-				var credentials = plugin.getAuthCredentials("NewPct", reason, do_query, 'newpct');
-				if(!credentials)
-					{
-					if(!do_query)
-						{
-						do_query = true;
-						continue;
-						}
-					break;
-					}
-				if (credentials.rejected) {break;}
-				if (credentials.username == "" || credentials.password == "")
-					{
-					if (!do_query)
-						{
-						do_query = true;
-						continue;
-						}
-					break;
-					}
-
-				try
-					{
-					var datos_post = {'userName': credentials.username,'userPass': credentials.password};
-					file_contents = post_urlsource('http://www.newpct1.com/entrar',datos_post);
-					var usuario = extraer_texto(file_contents,"loadtopBar('","'");
-					if(usuario!='')
-						{
-						if(do_query == true) {showtime.notify('Logueado en NewPct como: ' + credentials.username, 3);}
-						valor_retorno = get_urlsource(url_servidor);
-						break;
-						}
-					else
-						{
-						reason = "Usuario/Contraseña Incorrecta.";
-						continue;
-						}
-					}
-					catch (ex) { showtime.trace('NewPct: error al autenticarse.'); break; }
-				}
-			}
-		else {valor_retorno = file_contents;} //logueado
-
-		return valor_retorno;
-		}
-
-
 		this.parsenewpct= function (params) 
 		{	
 		/*var params={'url_servidor': ,'page_uri': ,'uri_siguiente': ,'subtitulo': }*/
-		var numero_pagina = parseInt(extraer_texto(params.url_servidor,'pag=','&'));			
+		
+		var numero_pagina =  parseInt(params.url_servidor.substr(params.url_servidor.lastIndexOf("pg=")+3));
+		
 		var file_contents = get_urlsource(params.url_servidor);
-		var ultima_pagina = extraer_texto(file_contents,"<div id='centPag'>","</div>");
-		ultima_pagina = ultima_pagina.substr(ultima_pagina.lastIndexOf("<a href='javascript:;'  class='todo'  onclick=\"orderCategory('todo','','"));
-		ultima_pagina = extraer_texto(ultima_pagina,"<a href='javascript:;'  class='todo'  onclick=\"orderCategory('todo','','","'");
+		var ultima_pagina = extraer_texto(file_contents,'<ul class="pagination">','</ul>');
+		ultima_pagina = ultima_pagina.substr(ultima_pagina.lastIndexOf("<a href="));
+		ultima_pagina = extraer_texto(ultima_pagina,'pg=','"');
+		if (params.buscar == true)
+		{
+			file_contents = extraer_texto(file_contents,'<ul class="buscar-list">','</ul>');
+		}else{
+			file_contents = extraer_texto(file_contents,'<ul class="pelilist">','</ul>');
+		}
+		
 		var array_aux = extraer_html_array(file_contents,'<li>','</li>');
 		file_contents = "";
 
@@ -1677,15 +1498,27 @@
 
 		for (var i=0;i<array_aux.length;i++)
 			{
-			imagen=extraer_texto(array_aux[i],"<img src='","'");
-			url_video=extraer_texto(array_aux[i],"<a href='","'>");
-			file_contents = extraer_texto(array_aux[i], "<div class='cover boxcaption'>", "</div>");
-		
-			titulo=extraer_texto(file_contents,'<h3>','</h3>');
-			if (params.subtitulo) titulo=titulo + ' ' + extraer_texto(array_aux[i],'<p>','<br/>');
-			//calidad = extraer_texto(file_contents, "Calidad:", "<br>");
-           // idioma = extraer_texto(file_contents, "Idioma :", "</p>");
-	
+			imagen=extraer_texto(array_aux[i],'<img src="','"');
+			if (params.serie==true)
+			{
+				url_video=extraer_texto(array_aux[i],'<a href="','"');
+			}else{
+				url_video='http://www.newpct1.com/ver-online/' + extraer_texto(array_aux[i],'<a href="http://www.newpct1.com/','"');
+			}
+			
+			
+			if (params.buscar == true)
+			{
+				titulo=extraer_texto(array_aux[i],'<h2 style="padding:0;">','</h2>');
+			}else{
+				titulo=extraer_texto(array_aux[i],'<h2>','</h2>');
+			}
+			//Eliminar tag html
+			var regex = /(<([^>]+)>)/ig 
+			titulo=titulo.replace(regex,"");
+			
+			//if (params.subtitulo) titulo=titulo + ' ' + extraer_texto(array_aux[i],'<p>','<br/>');
+				
 			array_playlist.push(new Item_menu(titulo,imagen,params.page_uri,url_video));
 			}
 
@@ -1693,7 +1526,7 @@
 		var pagina_siguiente = (parseInt(numero_pagina) + 1);
 		if(numero_pagina<ultima_pagina)
 			{
-				array_playlist.push(new Item_menu('Siguiente',"views/img/siguiente.png",params.uri_siguiente,params.url_servidor.replace('pag=' + numero_pagina,'pag=' + pagina_siguiente)));		
+				array_playlist.push(new Item_menu('Siguiente',"views/img/siguiente.png",params.uri_siguiente,params.url_servidor.replace('pg=' + numero_pagina,'pg=' + pagina_siguiente)));		
 			}
 
 		return array_playlist;
@@ -1710,21 +1543,25 @@
 
 		if(texto_busqueda != undefined)
 			{
-			var datos_post = 
+			/*var datos_post = 
 				{
-				'cID': 0, 'tLang': 0, 'oBy': 0,	'oMode': 0, 'category_': op_categoria, 'subcategory_' : 'All',
+				'cID': 0, 'tLang': 0, 'oBy': 0,	'oMode': 0, 'categoryIDR': op_categoria, 'subcategory_' : 'All',
 				'idioma_': 'All', 'calidad_': 'All', 'oByAux': 0, 'oModeAux': 0, 'size_': 0,
 				'q': texto_busqueda, 'btnb': 'Filtrar+Busqueda'
-				};
+				}; // http://www.newpct1.com/index.php?page=buscar&url=&letter=&q=pedro&categoryID=&categoryIDR=757&calidad=&idioma=&ordenar=Fecha&inon=Descendente&pg=1
 			var file_contents = post_urlsource(url_servidor,datos_post);
-			var resultados = file_contents.indexOf('No hemos encontrado resultados');
-
-			page.metadata.title = page.metadata.title + texto_busqueda;
+			var aux_string = extraer_texto(file_contents ,'<ul class="buscar-list">','</ul>');
+			var array_aux = extraer_html_array(aux_string,'<li>','</li>');
+			
+			
+			
+			
 
 			if(resultados==-1)
 				{
-				var aux_string = extraer_texto(file_contents ,'<tbody>','</tbody>');
-				var array_aux = extraer_html_array(aux_string,'<tr>','</tr>');
+
+				showtime.print(aux_string)				
+				
 				array_aux.sort();
 
 				var titulo;
@@ -1742,22 +1579,23 @@
 						array_playlist.push(new Item_menu(titulo,imagen,page_uri,url_video));
 						}
 					}
-				}
-			//devuelve bastantes resultados no veo necesario usar paginas
+				}*/
+			
+			// http://www.newpct1.com/index.php?page=buscar&url=&letter=&q=pedro&categoryID=&categoryIDR=757&calidad=&idioma=&ordenar=Fecha&inon=Descendente&pg=1
+			page.metadata.title = page.metadata.title + texto_busqueda;
+			
+			var url='http://www.newpct1.com/index.php?page=buscar&url=&letter=&q=' + texto_busqueda + '&categoryID=&categoryIDR=' + op_categoria + '&calidad=&idioma=&ordenar=Fecha&inon=Descendente&pg=1';
+			
+			var params={'url_servidor': unescape(url),
+					'page_uri': ':verenlaces:newpct:',
+					'uri_siguiente': ':vercontenido:newpct:tipo1:',
+					'subtitulo':false,
+					'buscar':true}	
+			
+			array_playlist= that.parsenewpct (params);
+			
+			
 			}
-
-		return array_playlist;
-		}
-
-
-		this.parsenewpcttipodefault = function(url_servidor,params){
-			url_servidor=unescape(url_servidor);
-			var array_playlist=[];
-
-			var file_contents = get_urlsource(url_servidor);
-			//check_login
-			file_contents = that.checkloginnewpct(url_servidor, file_contents);
-			if(file_contents!=false) array_playlist= this.parsenewpct (params);
 
 		return array_playlist;
 		}
@@ -2972,9 +2810,9 @@
 		this.getmenu= function(page){
 		//retorna el Menu
 			var array_menu=[
-				new Item_menu('Ultimos Capitulos','views/img/folder.png',':vercontenido:newpctseries:ultimoscapitulos:'+ escape('http://www.newpct1.com/series-alta-definicion-hd/')),
-				//new Item_menu('Orden Alfabetico','views/img/folder.png',':alfabeto:newpctseries:09'),
-				new Item_menu('Buscar Serie HD','views/img/search.png',':vercontenido:newpctseries:tipobusqueda:'+ escape('http://www.newpct1.com/buscar-descargas/'))
+				new Item_menu('Ultimos Capitulos','views/img/folder.png',':vercontenido:newpctseries:ultimoscapitulos:'+ escape('http://www.newpct1.com/index.php?page=categorias&url=series-hd/&letter=&pg=1')),
+				new Item_menu('Orden Alfabetico','views/img/folder.png',':alfabeto:newpctseries:0-9'),
+				new Item_menu('Buscar Serie HD','views/img/search.png',':vercontenido:newpctseries:tipobusqueda:'+ escape('http://www.newpct1.com/buscar'))
 				];
 		return array_menu;
 		}
@@ -2993,12 +2831,14 @@
 				{
 				case "ultimoscapitulos": 
 					page.metadata.title = "NewPct Series HD";
-					var url_2=escape('http://www.newpct1.com/include.inc/ajax.php/orderCategory.php?type=todo&leter=&sql=SELECT+DISTINCT+++%09%09%09%09%09%09torrentID%2C+++%09%09%09%09%09%09torrentCategoryID%2C+++%09%09%09%09%09%09torrentCategoryIDR%2C+++%09%09%09%09%09%09torrentImageID%2C+++%09%09%09%09%09%09torrentName%2C+++%09%09%09%09%09%09guid%2C+++%09%09%09%09%09%09torrentShortName%2C++%09%09%09%09%09%09torrentLanguage%2C++%09%09%09%09%09%09torrentSize%2C++%09%09%09%09%09%09calidad+as+calidad_%2C++%09%09%09%09%09%09torrentDescription%2C++%09%09%09%09%09%09torrentViews%2C++%09%09%09%09%09%09rating%2C++%09%09%09%09%09%09n_votos%2C++%09%09%09%09%09%09vistas_hoy%2C++%09%09%09%09%09%09vistas_ayer%2C++%09%09%09%09%09%09vistas_semana%2C++%09%09%09%09%09%09vistas_mes++%09%09%09%09++FROM+torrentsFiles+as+t+WHERE++(torrentStatus+%3D+1+OR+torrentStatus+%3D+2)++AND+(torrentCategoryID+IN+(1772%2C+1582%2C+1473%2C+1708%2C+1474%2C+1603%2C+1596%2C+1611%2C+1693%2C+1699%2C+1759%2C+1769%2C+1598%2C+1514%2C+1605%2C+1585%2C+1472%2C+1754%2C+1689%2C+1475%2C+1687%2C+1649%2C+1643%2C+1476%2C+1486%2C+1618%2C+1490%2C+1657%2C+1606%2C+1498%2C+1493%2C+1639%2C+1488%2C+1684%2C+1505%2C+1691%2C+1495%2C+1624%2C+1470%2C+1746%2C+1676%2C+1629%2C+1511%2C+1748%2C+1677%2C+1484%2C+1485%2C+1580%2C+1763%2C+1744%2C+1481%2C+1520%2C+1696%2C+1492%2C+1508%2C+1727%2C+1711%2C+1579%2C+1489%2C+1706%2C+1757%2C+1487%2C+1583%2C+1477%2C+1701%2C+1518%2C+1526%2C+1654%2C+1694%2C+1491%2C+1478%2C+1681%2C+1714%2C+1668%2C+1619%2C+1581%2C+1479%2C+1483%2C+1500%2C+1729%2C+1584%2C+1740%2C+1602%2C+1646%2C+1656%2C+1471%2C+1469))++AND+home_active+%3D+0++++ORDER+BY+torrentDateAdded++DESC++LIMIT+0%2C+50&pag=1&tot=&ban=3&cate=1469');
-					var params={'url_servidor': unescape(url_2),
-								'page_uri': ':verenlaces:newpctseries:',
+					//var url_2=escape('http://www.newpct1.com/include.inc/ajax.php/orderCategory.php?type=todo&leter=&sql=SELECT+DISTINCT+++%09%09%09%09%09%09torrentID%2C+++%09%09%09%09%09%09torrentCategoryID%2C+++%09%09%09%09%09%09torrentCategoryIDR%2C+++%09%09%09%09%09%09torrentImageID%2C+++%09%09%09%09%09%09torrentName%2C+++%09%09%09%09%09%09guid%2C+++%09%09%09%09%09%09torrentShortName%2C++%09%09%09%09%09%09torrentLanguage%2C++%09%09%09%09%09%09torrentSize%2C++%09%09%09%09%09%09calidad+as+calidad_%2C++%09%09%09%09%09%09torrentDescription%2C++%09%09%09%09%09%09torrentViews%2C++%09%09%09%09%09%09rating%2C++%09%09%09%09%09%09n_votos%2C++%09%09%09%09%09%09vistas_hoy%2C++%09%09%09%09%09%09vistas_ayer%2C++%09%09%09%09%09%09vistas_semana%2C++%09%09%09%09%09%09vistas_mes++%09%09%09%09++FROM+torrentsFiles+as+t+WHERE++(torrentStatus+%3D+1+OR+torrentStatus+%3D+2)++AND+(torrentCategoryID+IN+(1772%2C+1582%2C+1473%2C+1708%2C+1474%2C+1603%2C+1596%2C+1611%2C+1693%2C+1699%2C+1759%2C+1769%2C+1598%2C+1514%2C+1605%2C+1585%2C+1472%2C+1754%2C+1689%2C+1475%2C+1687%2C+1649%2C+1643%2C+1476%2C+1486%2C+1618%2C+1490%2C+1657%2C+1606%2C+1498%2C+1493%2C+1639%2C+1488%2C+1684%2C+1505%2C+1691%2C+1495%2C+1624%2C+1470%2C+1746%2C+1676%2C+1629%2C+1511%2C+1748%2C+1677%2C+1484%2C+1485%2C+1580%2C+1763%2C+1744%2C+1481%2C+1520%2C+1696%2C+1492%2C+1508%2C+1727%2C+1711%2C+1579%2C+1489%2C+1706%2C+1757%2C+1487%2C+1583%2C+1477%2C+1701%2C+1518%2C+1526%2C+1654%2C+1694%2C+1491%2C+1478%2C+1681%2C+1714%2C+1668%2C+1619%2C+1581%2C+1479%2C+1483%2C+1500%2C+1729%2C+1584%2C+1740%2C+1602%2C+1646%2C+1656%2C+1471%2C+1469))++AND+home_active+%3D+0++++ORDER+BY+torrentDateAdded++DESC++LIMIT+0%2C+50&pag=1&tot=&ban=3&cate=1469');
+					var params={'url_servidor': unescape(url),
+								//'page_uri': ':verenlaces:newpctseries:',
+								'page_uri': ':vercontenido:newpctseries:getCapitulos:',
 								'uri_siguiente': ':vercontenido:newpctseries:tipo1:',
-								'subtitulo': true}
-					array_playlist=this.parsenewpcttipodefault(url,params);
+								'subtitulo': false,
+								'serie':true}
+					array_playlist= that.parsenewpct (params);
 					break;
 				case "tipobusqueda":
 					page.metadata.title = "NewPct Series HD - Buscar - "; 
@@ -3006,23 +2846,31 @@
 					break;
 				case "tipo1":
 					var params={'url_servidor': unescape(url),
-								'page_uri': ':verenlaces:newpctseries:',
+								'page_uri': ':vercontenido:newpctseries:getCapitulos:',
 								'uri_siguiente': ':vercontenido:newpctseries:tipo1:',
-								'subtitulo': true}
+								'subtitulo': false,
+								'serie':true}
 					array_playlist= that.parsenewpct (params);
 					break;
 				case "alfabeto":
 					page.metadata.title = "NewPct Series HD - Orden Alfabetico: " + ((url=='09')?'0-9': url.toUpperCase()); //en este caso url almacena la letra a buscar
-					var url_2=escape('http://www.newpct1.com/include.inc/ajax.php/orderCategory.php?type=letter&leter=' + url + '&sql=SELECT+DISTINCT+++%09%09%09%09%09%09torrentID%2C+++%09%09%09%09%09%09torrentCategoryID%2C+++%09%09%09%09%09%09torrentCategoryIDR%2C+++%09%09%09%09%09%09torrentImageID%2C+++%09%09%09%09%09%09torrentName%2C+++%09%09%09%09%09%09guid%2C+++%09%09%09%09%09%09torrentShortName%2C++%09%09%09%09%09%09torrentLanguage%2C++%09%09%09%09%09%09torrentSize%2C++%09%09%09%09%09%09calidad+as+calidad_%2C++%09%09%09%09%09%09torrentDescription%2C++%09%09%09%09%09%09torrentViews%2C++%09%09%09%09%09%09rating%2C++%09%09%09%09%09%09n_votos%2C++%09%09%09%09%09%09vistas_hoy%2C++%09%09%09%09%09%09vistas_ayer%2C++%09%09%09%09%09%09vistas_semana%2C++%09%09%09%09%09%09vistas_mes++%09%09%09%09++FROM+torrentsFiles+as+t+WHERE++(torrentStatus+%3D+1+OR+torrentStatus+%3D+2)++AND+(torrentCategoryID+IN+(1772%2C+1582%2C+1473%2C+1708%2C+1474%2C+1603%2C+1596%2C+1611%2C+1693%2C+1699%2C+1759%2C+1769%2C+1598%2C+1514%2C+1605%2C+1585%2C+1472%2C+1754%2C+1689%2C+1475%2C+1687%2C+1649%2C+1643%2C+1476%2C+1486%2C+1618%2C+1490%2C+1657%2C+1606%2C+1498%2C+1493%2C+1639%2C+1488%2C+1684%2C+1505%2C+1691%2C+1495%2C+1624%2C+1470%2C+1746%2C+1676%2C+1629%2C+1511%2C+1748%2C+1677%2C+1484%2C+1485%2C+1580%2C+1763%2C+1744%2C+1481%2C+1520%2C+1696%2C+1492%2C+1508%2C+1727%2C+1711%2C+1579%2C+1489%2C+1706%2C+1757%2C+1487%2C+1583%2C+1477%2C+1701%2C+1518%2C+1526%2C+1654%2C+1694%2C+1491%2C+1478%2C+1681%2C+1714%2C+1668%2C+1619%2C+1581%2C+1479%2C+1483%2C+1500%2C+1729%2C+1584%2C+1740%2C+1602%2C+1646%2C+1656%2C+1471%2C+1469))++AND+home_active+%3D+0++++ORDER+BY+torrentDateAdded++DESC++LIMIT+0%2C+50&pag=1&tot=&ban=3&cate=1469');
+					var url_2=escape('http://www.newpct1.com/series-hd/letter/' + url); 
 					var params={'url_servidor': unescape(url_2),
 								'page_uri': ':vercontenido:newpctseries:getCapitulos:',
-								'uri_siguiente': ':vercontenido:newpctseries:tipo2:', //el tipo2 no esta implementado por q no contemplamos paginacion aqui
-								'subtitulo': false}
+								'uri_siguiente': ':vercontenido:newpctseries:tipo1:', 
+								'subtitulo': false,
+								'serie':true}
 					array_playlist= that.parsenewpct (params);
 					break;
 				case "getCapitulos":
-					page.metadata.title = "NewPct Series HD";
-					array_playlist=parsenewpctseriesCapitulos(url);
+					page.metadata.title = "NewPct Series HD: Capitulos";
+					var params={'url_servidor': unescape(url),
+								'page_uri': ':verenlaces:newpctseries:',
+								//'page_uri': ':vercontenido:newpctseries:getCapitulos:',
+								'uri_siguiente': ':vercontenido:newpctseries:tipo1:',
+								'subtitulo': false,
+								'buscar':true}
+					array_playlist= that.parsenewpct (params);
 					break;
 				}
 		return array_playlist;
@@ -3030,77 +2878,7 @@
 
 
 
-		/************************************************************************************
-		/*	funcion getservidores: Devuelve un listado de enlaces al contenido en los 		*
-		/*							servidores soportados. 								    *
-		/*	Parametros: 																    *
-		/*		url: direccion de la que se debe extraer la lista.							*
-		/*       page: referencia a la pagina de showtime desde donde se llama a la funcion.*
-		/*	Retorna: Array de servidores												    *
-		/************************************************************************************/
-		this.getservidores= function (url,page){
-			var array_servidores=[];
-			url=unescape(url);
-			var file_contents = get_urlsource(url);
-
-			var titulo;
-			var imagen;
-			var url_host;
-			var servidor;
-			var idioma;
-			var calidad;
-			var descripcion;
-
-			//check login
-			file_contents = this.checkloginnewpct(url, file_contents);
-			if(file_contents!=false)
-			{
-				//item_Actual
-				imagen = extraer_texto(file_contents ,'<div id="left_ficha">','" />');
-				imagen = extraer_texto(imagen,"src='","'"); 
-				
-				titulo = extraer_texto(file_contents ,'<title>','</title>');
-				titulo = titulo.substr(0,titulo.indexOf('Descarga'));
-				
-				descripcion = extraer_texto(file_contents ,' itemprop="description"  >',"</div>"); 
-
-				this.item_Actual=new Item_menu(titulo,imagen,null,url,descripcion);
-
-				file_contents = extraer_texto(file_contents,"<thead id='ver-online'>","</table>");
-				file_contents = extraer_texto(file_contents,'<tbody>','</tbody>');
-
-				var array_aux = extraer_html_array(file_contents,'<tr','</tr>');
-				file_contents = "";
-
-				var array_aux2=[];
-				for (var i=0;i<array_aux.length;i++)
-					{
-					if(array_aux[i].indexOf('<td>ver en 1 Link</td>')!='-1')
-						{
-						url_host = extraer_texto(array_aux[i],"<a href='","'");
-						array_aux2 = extraer_html_array(array_aux[i],'<td>','</td>');
-						servidor = extraer_texto(array_aux2[0],'<td>','</td>');	
-						idioma = extraer_texto(array_aux2[1],'<td>','</td>');
-						calidad = extraer_texto(array_aux2[2],'<td>','</td>');
-
-						var params={
-							"url_host" : url_host,
-							"servidor" : servidor,
-							"idioma" : idioma,
-							"calidad" : calidad
-						};
-
-						var objHost=HostFactory.createHost(servidor,params)
-						if (objHost)
-							{ 
-								array_servidores.push(objHost);
-							} 
-						}
-					}
-
-			}
-		return array_servidores;
-		}
+		
 
 		/************************************************************************************
 		/*	funcion getitem_alfabeto: Devuelve un listado de las subsecciones del canal. 	*
@@ -6880,7 +6658,7 @@ function utf8_encode(argString) {
 	var licencia = '';
 	var licencia_md5 = '1e7c1eac9e131fe3a01d7594c071bf8c'; //not4kids
 
-	var canal_test=true; //for testing only
+	var canal_test=false; //for testing only
 
 	var objCanal;
 	var objHistorial=new StoreHistorial();